changed health and damage datatypes to int for cleaner numbers

+++ Assets/_Complete-Game/Scripts/Final Project/EnemyBuilder.cs
@@ -12,7 +12,7 @@ namespace Completed
         private Animator animator;
         private Transform target;
         private bool skipMove;
-        public float hp;
+        public int hp;
         public AudioClip chopSound1;                //1 of 2 audio clips that play when the enemy is attacked by the player.
         public AudioClip chopSound2;                //2 of 2 audio clips that play when the enemy is attacked by the player.
         public List<GameObject> currentLevelBosses = new List<GameObject>();
@@ -101,7 +101,7 @@ namespace Completed
         }
 
         //DamageEnemy is called when the player attacks a enemy.
-        public IEnumerator DamageEnemy(float loss)
+        public IEnumerator DamageEnemy(int loss)
         {
             //Call the RandomizeSfx function of SoundManager to play one of two chop sounds.
             SoundManager.instance.RandomizeSfx(chopSound1, chopSound2);
